code refactoring/clean up, updated libraries, fixed toast
