+++ package.json
@@ -15,50 +15,50 @@
 		"test:unit": "vitest"
 	},
 	"devDependencies": {
-		"@playwright/test": "^1.46.1",
-		"@sveltejs/adapter-auto": "^3.2.4",
-		"@sveltejs/kit": "^2.5.22",
-		"@sveltejs/vite-plugin-svelte": "^3.1.1",
+		"@playwright/test": "^1.48.0",
+		"@sveltejs/adapter-auto": "^3.2.5",
+		"@sveltejs/kit": "^2.7.0",
+		"@sveltejs/vite-plugin-svelte": "^3.1.2",
 		"@types/dotenv-flow": "^3.3.3",
-		"@types/eslint": "8.56.0",
-		"@typescript-eslint/eslint-plugin": "^6.21.0",
-		"@typescript-eslint/parser": "^6.21.0",
+		"@types/eslint": "^9.6.1",
+		"@typescript-eslint/eslint-plugin": "^8.8.1",
+		"@typescript-eslint/parser": "^8.8.1",
 		"autoprefixer": "^10.4.20",
 		"dotenv-flow": "^4.1.0",
 		"echarts": "^5.5.1",
-		"eslint": "^8.57.0",
+		"eslint": "^9.12.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.43.0",
-		"flowbite": "^2.5.1",
-		"flowbite-svelte": "^0.44.24",
-		"lucia": "^3.2.0",
-		"postcss": "^8.4.41",
-		"postcss-load-config": "^5.1.0",
+		"eslint-plugin-svelte": "^2.44.1",
+		"flowbite": "^2.5.2",
+		"flowbite-svelte": "^0.46.23",
+		"lucia": "^3.2.1",
+		"postcss": "^8.4.47",
+		"postcss-load-config": "^6.0.1",
 		"prettier": "^3.3.3",
-		"prettier-plugin-svelte": "^3.2.6",
-		"prettier-plugin-tailwindcss": "^0.5.14",
-		"svelte": "^4.2.18",
-		"svelte-check": "^3.8.5",
+		"prettier-plugin-svelte": "^3.2.7",
+		"prettier-plugin-tailwindcss": "^0.6.8",
+		"svelte": "^4.2.19",
+		"svelte-check": "^4.0.5",
 		"svelte-echarts": "^0.1.1",
-		"tailwindcss": "^3.4.10",
-		"tslib": "^2.6.3",
-		"typescript": "^5.5.4",
-		"vite": "^5.4.1",
+		"tailwindcss": "^3.4.13",
+		"tslib": "^2.7.0",
+		"typescript": "^5.6.3",
+		"vite": "^5.4.8",
 		"vitest": "^1.6.0"
 	},
 	"type": "module",
 	"dependencies": {
 		"@node-rs/argon2": "^1.8.3",
-		"@sveltejs/adapter-vercel": "^5.4.3",
+		"@sveltejs/adapter-vercel": "^5.4.5",
 		"arctic": "^1.9.2",
 		"dotenv": "^16.4.5",
-		"firebase": "^10.13.0",
+		"firebase": "^10.14.1",
 		"lucide-svelte": "^0.312.0",
 		"moment": "^2.30.1",
-		"moment-timezone": "^0.5.45",
+		"moment-timezone": "^0.5.46",
 		"oslo": "^1.2.1",
 		"qr-scanner": "^1.4.2",
 		"resend": "^4.0.0",
-		"tailwind-merge": "^2.5.2"
+		"tailwind-merge": "^2.5.3"
 	}
 }
