+++ src/components/charts/TicketsECharts.svelte
