+++ src/routes/api/tickets/[ticketID]/+server.ts
@@ -3,7 +3,7 @@ import { getClientDB } from '$lib/firebase/client.js';
 import type { Ticket } from '$models/ticket';
 import { convertCode } from '$lib/codeConverter';
 import type { User } from 'lucia';
-import { hasPermission } from '$lib/utils';
+import { hasPermission } from '$lib/utils/permissions';
 import { UserPermissions } from '$models/permissions';
 
 export async function GET( { params, locals } ) {
