+++ src/routes/api/tickets/checkout/[ticketID]/+server.ts
@@ -1,11 +1,11 @@
-import { getClientDB } from '$lib/firebase/client.js';
+import { getClientDB } from '$lib/firebase/client';
 
 import { Timestamp, collection, doc, getDoc, updateDoc } from 'firebase/firestore';
-import type { Ticket } from '$models/ticket.js';
-import { convertCode } from '$lib/codeConverter.js';
+import type { Ticket } from '$models/ticket';
+import { convertCode } from '$lib/codeConverter';
 import type { User } from 'lucia';
-import { hasPermission } from '$lib/utils.js';
-import { UserPermissions } from '$models/permissions.js';
+import { hasPermission } from '$lib/utils/permissions';
+import { UserPermissions } from '$models/permissions';
 
 export async function PUT( { params, locals } ) {
 	if(!locals.user){
