+++ src/routes/check-in/+page.server.ts
@@ -1,4 +1,4 @@
-import { hasPermission } from '$lib/utils';
+import { hasPermission } from '$lib/utils/permissions';
 import { redirect } from '@sveltejs/kit';
 import type { PageServerLoad } from '../$types';
 import { UserPermissions } from '$models/permissions';
