+++ src/routes/generate/+page.server.ts
@@ -2,7 +2,7 @@ import { getClientDB } from '$lib/firebase/client.js';
 import { fail, redirect } from '@sveltejs/kit';
 import { doc, setDoc } from 'firebase/firestore';
 import type { PageServerLoad } from '../$types';
-import { hasPermission } from '$lib/utils';
+import { hasPermission } from '$lib/utils/permissions';
 import { UserPermissions } from '$models/permissions';
