+++ src/routes/tickets/+page.server.ts
@@ -1,5 +1,5 @@
 import { getClientDB } from "$lib/firebase/client";
-import { hasPermission } from "$lib/utils";
+import { hasPermission } from "$lib/utils/permissions";
 import { redirect } from "@sveltejs/kit";
 import { collection, getDocs, query, where } from "firebase/firestore";
 import type { User } from "lucia";
